TempFileCleanup.bat sort user directory cleaning

+++ resources/stage_1_tempclean/tempfilecleanup/TempFileCleanup.bat
@@ -79,43 +79,43 @@ REM if exist %SystemDrive%\$Windows.~WS (
 :: so if we don't find "Microsoft" in the first 9 characters we can safely assume we're not on XP/2k3.
 if /i "%WIN_VER:~0,9%"=="Microsoft" (
 	for /D %%x in ("%SystemDrive%\Documents and Settings\*") do (
-		del /F /S /Q "%%x\Local Settings\Temp\*" 2>NUL
-		del /F /S /Q "%%x\Recent\*" 2>NUL
-		del /F /S /Q "%%x\Local Settings\Temporary Internet Files\*" 2>NUL
-		del /F /S /Q "%%x\Local Settings\Application Data\ApplicationHistory\*" 2>NUL
-		del /F /S /Q "%%x\My Documents\*.tmp" 2>NUL
+		del /F /S /Q "%%x\Application Data\Adobe\Flash Player\*" 2>NUL
+		del /F /S /Q "%%x\Application Data\Macromedia\Flash Player\*" 2>NUL
 		del /F /S /Q "%%x\Application Data\Microsoft\Windows\WER\ReportArchive\*" 2>NUL
 		del /F /S /Q "%%x\Application Data\Microsoft\Windows\WER\ReportQueue\*" 2>NUL
 		del /F /S /Q "%%x\Application Data\Sun\Java\*" 2>NUL
-		del /F /S /Q "%%x\Application Data\Adobe\Flash Player\*" 2>NUL
-		del /F /S /Q "%%x\Application Data\Macromedia\Flash Player\*" 2>NUL
-		del /F /S /Q "%%x\Local Settings\Application Data\Microsoft\Internet Explorer\Recovery\*" 2>NUL
+		del /F /S /Q "%%x\Local Settings\Application Data\ApplicationHistory\*" 2>NUL
 		del /F /S /Q "%%x\Local Settings\Application Data\Google\Chrome\User Data\Default\Cache\*" 2>NUL
-		del /F /S /Q "%%x\Local Settings\Application Data\Google\Chrome\User Data\Default\JumpListIcons\*" 2>NUL
 		del /F /S /Q "%%x\Local Settings\Application Data\Google\Chrome\User Data\Default\JumpListIconsOld\*" 2>NUL
+		del /F /S /Q "%%x\Local Settings\Application Data\Google\Chrome\User Data\Default\JumpListIcons\*" 2>NUL
 		del /F /S /Q "%%x\Local Settings\Application Data\Google\Chrome\User Data\Default\Local Storage\*" 2>NUL
 		del /F /S /Q "%%x\Local Settings\Application Data\Google\Chrome\User Data\Default\Media Cache\*" 2>NUL
+		del /F /S /Q "%%x\Local Settings\Application Data\Microsoft\Internet Explorer\Recovery\*" 2>NUL
+		del /F /S /Q "%%x\Local Settings\Temp\*" 2>NUL
+		del /F /S /Q "%%x\Local Settings\Temporary Internet Files\*" 2>NUL
+		del /F /S /Q "%%x\My Documents\*.tmp" 2>NUL
+		del /F /S /Q "%%x\Recent\*" 2>NUL
 	)
 ) else (
 	for /D %%x in ("%SystemDrive%\Users\*") do ( 
-		del /F /S /Q "%%x\AppData\Local\Temp\*" 2>NUL
-		del /F /S /Q "%%x\AppData\Roaming\Microsoft\Windows\Recent\*" 2>NUL
-		del /F /S /Q "%%x\AppData\Local\Microsoft\Windows\Temporary Internet Files\*" 2>NUL
-		del /F /S /Q "%%x\AppData\Local\Microsoft\Windows\INetCache\*" 2>NUL
-		del /F /S /Q "%%x\My Documents\*.tmp" 2>NUL
-		del /F /S /Q "%%x\AppData\Local\Microsoft\Windows\WER\ReportArchive\*" 2>NUL
-		del /F /S /Q "%%x\AppData\Local\Microsoft\Windows\WER\ReportQueue\*" 2>NUL
-		del /F /S /Q "%%x\AppData\LocalLow\Sun\Java\*" 2>NUL
-		del /F /S /Q "%%x\AppData\Roaming\Adobe\Flash Player\*" 2>NUL
-		del /F /S /Q "%%x\AppData\Roaming\Macromedia\Flash Player\*" 2>NUL
-		del /F /S /Q "%%x\AppData\Local\Microsoft\Internet Explorer\Recovery\*" 2>NUL
-		del /F /S /Q "%%x\AppData\Local\Google\Chrome\User Data\Default\Cache\*" 2>NUL
-		del /F /S /Q "%%x\AppData\Local\Google\Chrome\User Data\Default\JumpListIcons\*" 2>NUL
-		del /F /S /Q "%%x\AppData\Local\Google\Chrome\User Data\Default\JumpListIconsOld\*" 2>NUL
-		del /F /S /Q "%%x\AppData\Local\Google\Chrome\User Data\Default\Local Storage\*" 2>NUL
-		del /F /S /Q "%%x\AppData\Local\Google\Chrome\User Data\Default\Media Cache\*" 2>NUL
 		del /F /S /Q "%%x\*.blf" 2>NUL
 		del /F /S /Q "%%x\*.regtrans-ms" 2>NUL
+		del /F /S /Q "%%x\AppData\LocalLow\Sun\Java\*" 2>NUL
+		del /F /S /Q "%%x\AppData\Local\Google\Chrome\User Data\Default\Cache\*" 2>NUL
+		del /F /S /Q "%%x\AppData\Local\Google\Chrome\User Data\Default\JumpListIconsOld\*" 2>NUL
+		del /F /S /Q "%%x\AppData\Local\Google\Chrome\User Data\Default\JumpListIcons\*" 2>NUL
+		del /F /S /Q "%%x\AppData\Local\Google\Chrome\User Data\Default\Local Storage\*" 2>NUL
+		del /F /S /Q "%%x\AppData\Local\Google\Chrome\User Data\Default\Media Cache\*" 2>NUL
+		del /F /S /Q "%%x\AppData\Local\Microsoft\Internet Explorer\Recovery\*" 2>NUL
+		del /F /S /Q "%%x\AppData\Local\Microsoft\Windows\INetCache\*" 2>NUL
+		del /F /S /Q "%%x\AppData\Local\Microsoft\Windows\Temporary Internet Files\*" 2>NUL
+		del /F /S /Q "%%x\AppData\Local\Microsoft\Windows\WER\ReportArchive\*" 2>NUL
+		del /F /S /Q "%%x\AppData\Local\Microsoft\Windows\WER\ReportQueue\*" 2>NUL
+		del /F /S /Q "%%x\AppData\Local\Temp\*" 2>NUL
+		del /F /S /Q "%%x\AppData\Roaming\Adobe\Flash Player\*" 2>NUL
+		del /F /S /Q "%%x\AppData\Roaming\Macromedia\Flash Player\*" 2>NUL
+		del /F /S /Q "%%x\AppData\Roaming\Microsoft\Windows\Recent\*" 2>NUL
+		del /F /S /Q "%%x\My Documents\*.tmp" 2>NUL
 	)
 )
